Refactor the security package by moving it to base package.
the security package was not in it's right location (by mistake) now
it's in it's suitable location
